+++ tests/test_loader.py
@@ -9,11 +9,11 @@
 """
 
 import pytest
+from marshmallow import fields
+
 from cfg_loader.exceptions import ConfigFileMissingError, ConfigFileNotFoundError
 from cfg_loader.loader import BaseConfigLoader, YamlConfigLoader
 from cfg_loader.schema import ConfigSchema
-from marshmallow import fields
-
 from .conftest import BASE_CONFIG_PATH
 
 
@@ -93,6 +93,24 @@ def test_yaml_config_loader_no_path(config_path):
     with pytest.raises(ConfigFileMissingError) as e:
         config_loader.load()
     assert 'VARIABLE' in str(e.value)
+
+
+@pytest.mark.parametrize('env_vars', [[('VARIABLE', None)]])
+def test_yaml_config_loader_with_default_path(config_path):
+    config_loader = YamlConfigLoader(ConfigSchemaTest,
+                                     substitution_mapping={'PATH': 'folder/file', 'SECRET': 'my-secret'},
+                                     config_file_env_var='VARIABLE',
+                                     default_config_path=BASE_CONFIG_PATH)
+
+    assert config_loader.load() == {
+        'base': {
+            'name': 'App-Name',
+            'path': '/home/user/folder/file',
+        },
+        'security': {
+            'secret': 'my-secret',
+        },
+    }
 
 
 def test_yaml_config_loader_invalid_path():